Move these two types to the more-general util package
