+++ src/main/java/org/wildfly/security/auth/util/PrincipalDecoder.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.wildfly.security.auth.spi;
+package org.wildfly.security.auth.util;
 
 import java.security.Principal;
