+++ src/main/java/org/wildfly/security/x500/X509CertificateCredentialDecoder.java
@@ -22,7 +22,7 @@ import java.security.cert.X509Certificate;
 
 import javax.security.auth.x500.X500Principal;
 
-import org.wildfly.security.auth.spi.CredentialDecoder;
+import org.wildfly.security.auth.util.CredentialDecoder;
 
 /**
  * A credential decoder which can decode an {@link X509Certificate} into an {@link X500Principal}.
